Add AssertPromise trait
